requirements.txt
langchain==0.3.26
langchain-openai==0.3.27
requests==2.32.4 
python-dotenv==1.0.1
langgraph==0.3.1